% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphx-package.R
\docType{package}
\name{graphx}
\alias{graphx}
\title{Graphics Routines for Scientific Research}
\description{
Routines to produce and export publication-ready scientific plots and movies (e.g., used in Obreschkow et al. (2020) <doi:10.1093/mnras/staa445>). These include special color scales, projection routines, and bitmap handling routines.
}
\author{
Danail Obreschkow
}
